[FEAT] 시장가 거래

+++ src/main/java/com/konkuk/kubit/utils/TransactionUtil.java
@@ -2,7 +2,7 @@ package com.konkuk.kubit.utils;
 
 public class TransactionUtil {
     public static final double CHARGE = 0.05;
-    public static double getCharge(double requestPrice, double quantity){
-        return requestPrice * quantity * CHARGE;
+    public static int getCharge(double requestPrice, double quantity){
+        return (int) (requestPrice * quantity * CHARGE);
     }
 }
